LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicyviolation-block-image-expected.txt

CONSOLE MESSAGE: Refused to load the image 'http://127.0.0.1:8000/security/resources/abe.png' because it violates the following Content Security Policy directive: "img-src 'none'".

Check that a SecurityPolicyViolationEvent is fired upon blocking an image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.e.documentURI is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/securitypolicyviolation-block-image.html"
PASS window.e.referrer is ""
PASS window.e.blockedURI is "http://127.0.0.1:8000/security/resources/abe.png"
PASS window.e.violatedDirective is "img-src 'none'"
PASS window.e.effectiveDirective is "img-src"
PASS window.e.originalPolicy is "img-src 'none'"
PASS window.e.sourceURL is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/securitypolicyviolation-block-image.html"
PASS window.e.lineNumber is 30
PASS successfullyParsed is true

TEST COMPLETE
